+++ src/Kiota.Builder/PathSegmenters/GoPathSegmenter.cs
@@ -6,7 +6,7 @@ using Kiota.Builder.CodeDOM;
 using Kiota.Builder.Extensions;
 
 namespace Kiota.Builder.PathSegmenters;
-public class GoPathSegmenter : CommonPathSegmenter
+public class GoPathSegmenter(string rootPath, string clientNamespaceName) : CommonPathSegmenter(rootPath, clientNamespaceName)
 {
     private static readonly HashSet<string> specialFileNameSuffixes = new(StringComparer.OrdinalIgnoreCase) {
         "test"       ,
@@ -69,7 +69,6 @@ public class GoPathSegmenter : CommonPathSegmenter
         "wasm"        ,
     };
 
-    public GoPathSegmenter(string rootPath, string clientNamespaceName) : base(rootPath, clientNamespaceName) { }
     public override string FileSuffix => ".go";
     public override IEnumerable<string> GetAdditionalSegment(CodeElement currentElement, string fileName)
     {
@@ -85,7 +84,13 @@ public class GoPathSegmenter : CommonPathSegmenter
         return currentElement switch
         {
             CodeNamespace => "go",
-            _ => GetLastFileNameSegment(currentElement).ToSnakeCase().EscapeSuffix(specialFileNameSuffixes).ShortenFileName(100),
+            _ => GetLastFileNameSegment(currentElement)
+                .Replace("_", "_escaped_", StringComparison.OrdinalIgnoreCase)// at this point we have the guarantee that the element name is unique based on how child elements are held in Kiota.
+                                                                              // The existence of the '_' character in an element name could make `ToSnakeCase` give the same file name.
+                                                                              // e.g. codeScanningVariantAnalysisStatus and codeScanningVariantAnalysis_status both give code_scanning_variant_analysis
+                                                                              // This shouldn't break anything as golang does not care about filenames
+                .ToSnakeCase()
+                .EscapeSuffix(specialFileNameSuffixes).ShortenFileName(100),
         };
     }
     public override string NormalizeNamespaceSegment(string segmentName) => segmentName?.ToLowerInvariant() ?? string.Empty;