+++ tests/Kiota.Builder.Tests/PathSegmenters/GoPathSegmenterTests.cs
@@ -6,11 +6,7 @@ namespace Kiota.Builder.Tests.PathSegmenters;
 
 public class GoPathSegmenterTests
 {
-    private readonly GoPathSegmenter segmenter;
-    public GoPathSegmenterTests()
-    {
-        segmenter = new GoPathSegmenter("D:\\source\\repos\\kiota-sample", "client");
-    }
+    private readonly GoPathSegmenter segmenter = new("D:\\source\\repos\\kiota-sample", "client");
 
     [Fact]
     public void GoPathSegmenterGeneratesCorrectFileName()
@@ -39,4 +35,21 @@ public class GoPathSegmenterTests
         var normalizedNamespace = segmenter.NormalizeNamespaceSegment(namespaceName);
         Assert.Equal("microsoft.graph", normalizedNamespace);// the file name should be lower case
     }
+
+    [Fact]
+    public void GoPathSegmenterGeneratesDistinctFileNames()
+    {
+        var fileName1 = segmenter.NormalizeFileName(new CodeClass
+        {
+            Name = "codeScanningVariantAnalysisStatus"
+        });
+        var fileName2 = segmenter.NormalizeFileName(new CodeClass
+        {
+            Name = "codeScanningVariantAnalysis_status"
+        });
+        Assert.NotEqual(fileName1, fileName2);// the file name should be Different!
+        Assert.Equal("code_scanning_variant_analysis_status", fileName1);// the file name should be camel cased!
+        Assert.Equal("code_scanning_variant_analysis_escaped_status", fileName2);// the file name should be camel cased!
+    }
+
 }